Bug #232: Disable connection pooling by default for SqlServer
Microsoft.Data.SqlClient v 5.x has connection pooling enabled by default. This can create issues for
grate due to that we expect connections to actually be closed when we close them. If they aren't
(which they aren't when connection pooling is enabled), some scripts, that alter the state of the database,
might fail.