Chore: Making the public API interface of grate more maintainable before exposing as a nuget package
When exposing grate as a nuget package, all of the source code is effectively a "user interface",
as opposed to when we only publish the command line artifact, where the command line is the only
user interface.

We don't want a huge API interface that users start to use and rely on, that we have to keep stable,
so I narrowed down the number of public types significantly (hope I didn't go too far).

In addition, I changed the unit tests to use the DI builders, so that we know that we support a lot
of common usage scenarios (which I hope the unit tests represent).

Some highlights:

* Use TryAddTransient and ..Singleton to avoid concurrency issues in dependency registration
* Removed ServiceCollection from GrateConfgurationBuilder
* Rewrote unit tests to use dependency injection - getting a feel on how it is to use