Add tests on the build, trimmed, self-contained command-line binaries
.Net gets more and more trimming and compiling features going forward, and, if we e.g. should replace System.Commandline, which
never seems to get out of beta, with another command-line parsing library, I feel much more confident doing this if we have a lot of tests
on the packaged an build binaries too, in addition to the DLLs directly.

I reused the existing unit tests, being a bit creative with the .csproj files and includes. I _hope_ it is a good idea, and that we don't have to
duplicate up any new tests we should write on the migrations for both DLLs/libraries and command-line. Including the files should make the
tests be available for both automagically.

* Add test projects for all databases for compiled command-line grate executable
* Add running of these tests to the Build and Release build pipeline
* Extracted unit test project imports to separate .targets file